fix: preserve daemon auth transport context
